requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.24
scikit-learn>=1.2
plotly>=5.15
openpyxl>=3.1
requests>=2.31
pillow>=10.0
